[FIX] add domain dependency for maven building

+++ domain/src/main/kotlin/app/domain/entities/BaseUser.kt
@@ -8,7 +8,7 @@ abstract class BaseUser {
     abstract var id: Long
     abstract var name: String
     abstract var lastName: String
-    abstract var phoneNumber: Int
+    abstract var phoneNumber: Long
     abstract var friendsIds: List<Long>
     abstract var avatarId: Long
     abstract var conversationsIds: List<Long>
